requirements.txt
instagrapi==1.17.13
python-dotenv
Pillow
chess
fentoboardimage
